Added ColorWheelTestBlock and removed ImageTestBlock
