+++ envs/dev/views/blocks/ColorWheelTestBlock.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * View file for block: ColorWheelTestBlock 
+ *
+ * File has been created with `block/create` command on LUYA version 1.0.0-dev. 
+ *
+ * @param $this->varValue('color');
+ *
+ * @var $this \luya\cms\base\PhpBlockView
+ */
+?>
