--- envs/dev/views/blocks/ImageTestBlock.php
@@ -1,8 +1,0 @@
-<?php use luya\lazyload\LazyLoad;
-
-?>
-<?= LazyLoad::widget([
-    'src' => $extras['image']->source,
-    'width' => $extras['image']->resolutionWidth,
-    'height' => $extras['image']->resolutionHeight,
-]) ?>
